requirements.txt
pytest~=7.4.4
pyperclip~=1.8.2
PyAutoGUI~=0.9.54
selenium~=4.16.0
requests~=2.31.0
python-dotenv~=1.0.1
pytest~=7.4.3
pytest-html==4.0.2
requests==2.31.0
webdriver_manager
allure-pytest
pytest-rerunfailures
pip
python-dotenv~=1.0.0
pyperclip~=1.8.2
boto3~=1.34.88
pynput ~=1.7.7
